5.2.0

+++ qtbase/tests/auto/network/access/qnetworkreply/tst_qnetworkreply.cpp
@@ -368,6 +368,8 @@ private Q_SLOTS:
 #ifdef QT_BUILD_INTERNAL
     void sslSessionSharing_data();
     void sslSessionSharing();
+    void sslSessionSharingFromPersistentSession_data();
+    void sslSessionSharingFromPersistentSession();
 #endif
 #endif
 
@@ -1414,7 +1416,8 @@ void tst_QNetworkReply::initTestCase()
 void tst_QNetworkReply::cleanupTestCase()
 {
 #if !defined Q_OS_WIN
-    QFile::remove(wronlyFileName);
+    if (!wronlyFileName.isNull())
+        QFile::remove(wronlyFileName);
 #endif
 #ifndef QT_NO_BEARERMANAGEMENT
     if (networkSession && networkSession->isOpen()) {
@@ -4387,6 +4390,10 @@ void tst_QNetworkReply::ioPostToHttpFromSocket_data()
 
 void tst_QNetworkReply::ioPostToHttpFromSocket()
 {
+    if (QTest::currentDataTag() == QByteArray("128k+1+proxyauth")
+            || QTest::currentDataTag() == QByteArray("128k+1+auth+proxyauth"))
+        QSKIP("Squid cannot handle authentication with POST data >= 64K (QTBUG-33180)");
+
     QFETCH(QByteArray, data);
     QFETCH(QUrl, url);
     QFETCH(QNetworkProxy, proxy);
@@ -5964,7 +5971,7 @@ void tst_QNetworkReply::sslSessionSharing()
     warmupRequest.setAttribute(QNetworkRequest::User, sessionSharingEnabled); // so we can read it from the slot
     if (! sessionSharingEnabled) {
         QSslConfiguration configuration(QSslConfiguration::defaultConfiguration());
-        configuration.setSslOption(QSsl::SslOptionDisableSessionTickets, true);
+        configuration.setSslOption(QSsl::SslOptionDisableSessionSharing, true);
         warmupRequest.setSslConfiguration(configuration);
     }
     QNetworkReply *reply = manager.get(warmupRequest);
@@ -6008,6 +6015,65 @@ void tst_QNetworkReply::sslSessionSharingHelperSlot()
         count = 0; // reset for next row
         sslSessionSharingWasUsed = false; // reset for next row
     }
+}
+
+void tst_QNetworkReply::sslSessionSharingFromPersistentSession_data()
+{
+    QTest::addColumn<bool>("sessionPersistenceEnabled");
+    QTest::newRow("enabled") << true;
+    QTest::newRow("disabled") << false;
+}
+
+void tst_QNetworkReply::sslSessionSharingFromPersistentSession()
+{
+    QString urlString("https://" + QtNetworkSettings::serverName());
+
+    // warm up SSL session cache to get a working session
+    QNetworkRequest warmupRequest(urlString);
+    QFETCH(bool, sessionPersistenceEnabled);
+    if (sessionPersistenceEnabled) {
+        QSslConfiguration warmupConfiguration(QSslConfiguration::defaultConfiguration());
+        warmupConfiguration.setSslOption(QSsl::SslOptionDisableSessionPersistence, false);
+        warmupRequest.setSslConfiguration(warmupConfiguration);
+    }
+    QNetworkReply *warmupReply = manager.get(warmupRequest);
+    warmupReply->ignoreSslErrors();
+    connect(warmupReply, SIGNAL(finished()), &QTestEventLoop::instance(), SLOT(exitLoop()));
+    QTestEventLoop::instance().enterLoop(20);
+    QVERIFY(!QTestEventLoop::instance().timeout());
+    QCOMPARE(warmupReply->error(), QNetworkReply::NoError);
+    QByteArray sslSession = warmupReply->sslConfiguration().sessionTicket();
+    QCOMPARE(!sslSession.isEmpty(), sessionPersistenceEnabled);
+
+    // test server sends a life time hint of 0 (old server) or 300 (new server),
+    // without session ticket we get -1
+    QList<int> expectedSessionTicketLifeTimeHint = sessionPersistenceEnabled
+            ? QList<int>() << 0 << 300 : QList<int>() << -1;
+    QVERIFY2(expectedSessionTicketLifeTimeHint.contains(
+                 warmupReply->sslConfiguration().sessionTicketLifeTimeHint()),
+             "server did not send expected session life time hint");
+
+    warmupReply->deleteLater();
+
+    // now send another request with a new QNAM and the persisted session,
+    // to verify it can be resumed without any internal state
+    QNetworkRequest request(warmupRequest);
+    if (sessionPersistenceEnabled) {
+        QSslConfiguration configuration = request.sslConfiguration();
+        configuration.setSessionTicket(sslSession);
+        request.setSslConfiguration(configuration);
+    }
+    QNetworkAccessManager newManager;
+    QNetworkReply *reply = newManager.get(request);
+    reply->ignoreSslErrors();
+    connect(reply, SIGNAL(finished()), &QTestEventLoop::instance(), SLOT(exitLoop()));
+    QTestEventLoop::instance().enterLoop(20);
+    QVERIFY(!QTestEventLoop::instance().timeout());
+    QCOMPARE(reply->error(), QNetworkReply::NoError);
+
+    bool sslSessionSharingWasUsedInReply = QSslConfigurationPrivate::peerSessionWasShared(
+                reply->sslConfiguration());
+    QCOMPARE(sessionPersistenceEnabled, sslSessionSharingWasUsedInReply);
 }
 
 #endif // QT_BUILD_INTERNAL
@@ -7522,7 +7588,7 @@ void tst_QNetworkReply::backgroundRequestInterruption()
         QNetworkSessionPrivate::setUsagePolicies(*const_cast<QNetworkSession *>(session.data()), original);
 
     QVERIFY(reply->isFinished());
-#ifdef Q_OS_MACX
+#ifdef Q_OS_OSX
     if (QSysInfo::MacintoshVersion == QSysInfo::MV_10_8)
         QEXPECT_FAIL("ftp, bg, nobg", "See QTBUG-32435", Abort);
 #endif
